requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.5
mysqlclient==2.2.7
passlib==1.7.4
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
